requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.22
pydantic==2.12.5
scikit-learn==1.7.2
pandas==2.3.3
numpy==2.2.6
